--- a/Sources/SwordFoundation/Scope.swift
+++ b/Sources/SwordFoundation/Scope.swift
@@ -1,4 +1,4 @@
-public enum Scope: String, Codable {
+public enum Scope: String, Codable, Sendable {
   case single
   case weakReference
 }