--- a/Sources/SwordGenerator/Factory/Factory.swift
+++ b/Sources/SwordGenerator/Factory/Factory.swift
@@ -1,0 +1,16 @@
+protocol Factory {
+  associatedtype T
+  func make() async -> FactoryResult<T>
+}
+
+enum FactoryResult<Success: Sendable>: Sendable {
+  case success(Success)
+  case failure([Report])
+
+  var reports: [Report] {
+    switch self {
+    case .success: []
+    case .failure(let reports): reports
+    }
+  }
+}