--- a/Sources/SwordGenerator/Model/ComponentTree.swift
+++ b/Sources/SwordGenerator/Model/ComponentTree.swift
@@ -1,0 +1,4 @@
+struct ComponentTree {
+  let rootComponent: Component
+  let subcomponentsByParent: [ComponentName: [Component]]
+}