--- a/Sources/SwordGenerator/Model/Module.swift
+++ b/Sources/SwordGenerator/Model/Module.swift
@@ -2,5 +2,6 @@ import Foundation
 
 struct Module {
   let name: String
+  let componentName: ComponentName
   let providers: [Provider]
 }