--- a/Sources/SwordGenerator/Model/Parameter.swift
+++ b/Sources/SwordGenerator/Model/Parameter.swift
@@ -2,7 +2,7 @@ import Foundation
 import SwiftSyntax
 import SwordFoundation
 
-struct Parameter: Codable {
+struct Parameter: Hashable, Codable {
   let key: Key
   let type: Type
   let name: String