… Sources/SwordGenerator/Descriptor/RootComponentDescriptor.swift
… Sources/SwordGenerator/Descriptor/SubcomponentDescriptor.swift
… Sources/SwordGenerator/Parser/Factory/BindingGraphFactory.swift
… Sources/SwordGenerator/Parser/Registry/ComponentRegistry.swift
… Sources/SwordGenerator/Parser/Registry/DependencyRegistry.swift
… Sources/SwordGenerator/Parser/Registry/ImportRegistry.swift
… Sources/SwordGenerator/Parser/Registry/ModuleRegistry.swift
… Sources/SwordGenerator/Parser/SwordParser.swift
… Sources/SwordGenerator/Parser/Validator/BindingGraphValidator.swift
… Sources/SwordGenerator/Parser/Validator/ComponentValidator.swift
… Sources/SwordGenerator/Parser/Validator/DependencyValidator.swift
… Sources/SwordGenerator/Parser/Validator/ModuleValidator.swift
… Sources/SwordGenerator/Parser/Validator/ValidationResult.swift
… Sources/SwordGenerator/Parser/Visitor/DependencyVisitor.swift
… Sources/SwordGenerator/Parser/Visitor/ImportVisitor.swift
… Sources/SwordGenerator/Parser/Visitor/ModuleVisitor.swift
… Sources/SwordGenerator/Parser/Visitor/SourceFileVisitor.swift
… Sources/SwordGenerator/Parser/Visitor/SubcomponentVisitor.swift
… Sources/SwordGenerator/SwordGenerator.swift