## Instruments measurement result

📝 with the same conditions as https://github.com/rockname/sword/pull/23

Total: 16.70 seconds (1.34 times faster than before)

| Name | Duration |
|--------|--------|
| parseSourceFiles | 152.58ms |
| makeComponentTree | 76.79μs |
| makeDependencies | 9.60ms |
| makeModules | 38.93ms |
| makeBindingGraph | 12.68s | 
| exportBindingGraph | 1.30s |